90% of offers  task

+++ lib/src/modules/offers/domain_layer/entities/coupon.dart
@@ -2,9 +2,9 @@ import 'package:equatable/equatable.dart';
 
 class Coupon extends Equatable {
   String text;
-  double discount;
-  DateTime date  ;
-  Coupon({required this.text, required this.discount , required this.date});
+  int discount;
+  DateTime date;
+  Coupon({required this.text, required this.discount, required this.date});
   @override
-  List<Object?> get props => [text, discount , date];
+  List<Object?> get props => [text, discount, date];
 }
